added fs_append_nb

+++ includes/libft_comparison.h
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   libft_other.h                                      :+:      :+:    :+:   */
+/*   libft_comparison.h                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ugwentzi <ugwentzi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/14 16:58:18 by ugwentzi          #+#    #+#             */
-/*   Updated: 2025/04/08 09:40:36 by ugwentzi         ###   ########.fr       */
+/*   Updated: 2025/04/14 15:27:33 by ugwentzi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
   │ comparison/ │
   └─────────────┘*/
 
-#include <stdbool.h>
+# include <stdbool.h>
 
 /**
  * @brief return the true/false value depending on `condition`.
@@ -35,6 +35,6 @@ void	*ft_ternary(bool condition, void *true_value, void *false_value);
  * @param true_value The value returned if condtion == true.
  * @param false_value The value returned if condition == false.
  */
-int		ft_iternary(bool condition, int true_value, int false_value);
+long	ft_iternary(bool condition, long true_value, long false_value);
 
 #endif
