+++ srcs/comparison/ft_iternary.c
@@ -6,13 +6,13 @@
 /*   By: ugwentzi <ugwentzi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/10 21:48:54 by ugwentzi          #+#    #+#             */
-/*   Updated: 2025/04/10 21:52:36 by ugwentzi         ###   ########.fr       */
+/*   Updated: 2025/04/14 14:32:18 by ugwentzi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_iternary(bool condition, int true_value, int false_value)
+long	ft_iternary(bool condition, long true_value, long false_value)
 {
 	if (condition)
 		return (true_value);
